+++ app/src/main/java/at/derfl007/jokinghazard/drag_and_drop/DragShadow.java
@@ -1,4 +1,4 @@
-package at.derfl007.jokinghazard.DragAndDrop;
+package at.derfl007.jokinghazard.drag_and_drop;
 
 import android.graphics.Canvas;
 import android.graphics.Color;
@@ -7,16 +7,17 @@ import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
 import android.view.View;
 
-public class DragShadow extends View.DragShadowBuilder{
+public class DragShadow extends View.DragShadowBuilder {
 
     private static Drawable shadow;
 
     public void MyDragShadowBuilder(View v) {
 
-    shadow = new ColorDrawable(Color.LTGRAY);
-}
+        shadow = new ColorDrawable(Color.LTGRAY);
+    }
+
     @Override
-    public void onProvideShadowMetrics (Point size, Point touch) {
+    public void onProvideShadowMetrics(Point size, Point touch) {
         int width, height;
         width = getView().getWidth() / 2;
         height = getView().getHeight() / 2;
@@ -29,5 +30,5 @@ public class DragShadow extends View.DragShadowBuilder{
     public void onDrawShadow(Canvas canvas) {
         shadow.draw(canvas);
     }
-    }
+}
