+++ app/src/main/java/at/derfl007/jokinghazard/fragments/GameBoardFragment.java
@@ -2,18 +2,13 @@ package at.derfl007.jokinghazard.fragments;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 
-import android.content.ClipData;
-import android.content.ClipDescription;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
 
-import at.derfl007.jokinghazard.DragAndDrop.DragListener;
 import at.derfl007.jokinghazard.R;
 import at.derfl007.jokinghazard.activities.MainActivity;
 import io.socket.client.Socket;
